just need to finish edit and challenge

+++ forms.py
@@ -1,9 +1,9 @@
-from wtforms import Form, StringField, TextAreaField, IntegerField, FloatField
+from wtforms import Form, StringField, TextAreaField, IntegerField, DecimalField
 
 class AddForm(Form):
     name = StringField("name")
     inventory = IntegerField("inventory")
-    price = FloatField("price")
+    price = DecimalField("price")
     description = TextAreaField("description")
 
 class DelForm(Form):
